Shorter XML production
